requirements.txt
torch>=2.0.0
transformers>=4.30.0
pyserini>=0.22.0
sentence-transformers>=2.2.2
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
tqdm>=4.65.0
faiss-cpu>=1.7.4
matplotlib>=3.7.0
cohere>=4.0.0
voyageai>=0.1.0
